refactor: move ai-query-generator into ensadmin
