+++ apps/ensadmin/src/components/ai-query-generator/anthropic-adapter.ts
