+++ apps/ensadmin/src/components/ai-query-generator/index.ts
