+++ apps/ensadmin/src/components/ai-query-generator/query-generator-client.ts
