+++ packages/ensnode-utils/package.json
@@ -51,10 +51,5 @@
     "typescript": "catalog:",
     "viem": "catalog:",
     "vitest": "catalog:"
-  },
-  "dependencies": {
-    "@anthropic-ai/sdk": "^0.39.0",
-    "@gqlpt/adapter-base": "0.0.0-alpha.31",
-    "gqlpt": "0.0.0-alpha.31"
   }
 }
